requirements.txt
flask
pandas
scikit-learn
joblib
psycopg2-binary
werkzeug
